Change order of rotation to yaw and then pitch
THREE is using local axis to rotate using Euler and the order is 'XYZ'
by default. Before this change, it would rotate around X first for
"pitch" which would make local Y is than world Y, once it rotates around
local "Y" for yaw the hotspot would end up being in the wrong place.
This is not visible in the hotspots example as no pitch is set on any of
the hotspots.